+++ lib/feature/catalog/ui/widgets/check_out_card.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../common/default_button.dart';
-import '../../../../common/model/order_model.dart';
-import '../../../../constants/products_consants.dart';
 import '../../../../constants/size_config.dart';
-import '../../../auth/service/user_service.dart';
 import '../../../order/ui/order_screen.dart';
 import '../../model/product.dart';
 
 class CheckoutCard extends StatelessWidget {
   final List<Product> productList;
-  final double totalAmount;
+  final int totalAmount;
   const CheckoutCard({
     Key? key,
     required this.totalAmount,
